LocalNode is now a general purpose server for all algorithms
Additionally, LocalNode contains a worker attribute which holds
a reference to a specific algorithm worker, each time an algorithm
is run.
For now, LocalNodes object still functions as a proxy for all available
workers.
Future work: implement iterator protocol for LocalNodes in order to
provide a proxy to each worker individually.