Worker is dynamically assigned in LocalNode at every call
This will allow us to run multiple algorithm tasks concurrently.
To achieve this, a task parameter is passed at every method call
to a local node. Currenlty the local node loads data from the DB
at every call, but in the future local methods will be replaced by
UDFs executed in the DB.