Add possibility to send expressions to be evaluated to workers
New feature: master now can use some mock DesignMatrix objects
which register computations and send the corresponding expressions
to workers. Workers can then evaluate this expressions using python's
eval. Some additional information is sent as well, specifically,
the names of the columns to which each desing matrix corresponds.

This eliminates the need to implement a separate class for algorithm
workers. All buisness logic can be contained in a single class. The
part of the computation that needs to be done in workers is transfered
to them as a simple string representing the expression to be evaluated.

This idea can be pursued further in two directions:

A) By parsing expressions and translating to opcodes, we can
probably determine automatically which part of the computation must be
performed in the workers and which in the master. This will eliminate
the need to write `self.workers.eval(expr)` and the algorithm logic
can be written agnostically of the worker-master structure. The key
idea in determining the centralized/decentralized parts of the
computation is the commutativity of matrix concatenation with other
operations, e.g. matrix multiplication. For example:
concat_i sum_j A_ij B_jk == sum_j concat_i A_ij B_jk.

B) A second point concerns privacy, specifically the local aggregates
question. Again by parsing and translating expressions to opcodes we can
obtain information on which operations result in aggregates and which
not by looking to the shape of the resulting tensor. Tensors (desing
matrices for the moment) have a shape which can hold a special
place-holder which represents the (unknown to the master) number of
rows. We can addopt as a general rule the fact that workers cannot send
results whose shape contain this special place-holder. That would mean
sending a number of quantities directly related to the number of rows,
hence from which inference in single data-points can be made.

Future work: determine the rules of "federated" algebra (see A).
Automate privacy checks.