Add wrapper for numpy arrays which can interact with mocks
Not all computation is done on workers, some must take place on master.
Thus we need two kind of variable objects.
One completely empty of data, which holds syntactic information only
and which must be evaluated on workers. And another which wraps
numpy arrays but also inherits from Matrix(Mock) and knows how
to interact with the first kind.

This second kind performs two
duties. First, it holds the actual data and registers them into an
array_register which, together with mock_register, is transfered to
local nodes. Second, it simulates the behaviour of mocks and thus
allows syntactic information to be carried further into the computation.