First working prototype of autoprivacy algorithm
Autoprivacy algorithm is implemented as a prototype and applied
to linear regression with success.

The algorithm purpose is to evaluate python expressions (for the time
being Linear Algebra expressions) on a privacy-preserving federated
system. The computation of each complex expression needs to be broken
down to local and global parts where local parts are executed
on the nodes of the federation where the data resides (and from where
it is not allowed to escape) and the global part is executed in some
special node (called central, master or simply global node) where
no data resides beyond what is sent by the local nodes.

Our definition of "privacy" (not true privacy) is that the local
nodes are only allowed to send *aggregate results* to the global
node. We define an aggregate result as the result of some kind of
reduction (contraction) over the length of the local datasets.
More precisely, if n is the length of the dataset,
an operation that takes an array of length *O(n)*,
along some dimension and outputs an array of length
*O(1)* across this same dimension.

The algorithm operates as follows:

First the expression is typed on the global node.
This expression involves some arrays whose content is not known
to the global node but whose shape is (shape is the length of each
dimension). The expression is parsed into an AST and then translated into
opcodes, using a specially crafted instruction set.
The instruction set is of the zero-address-format since it is designed
to operate on a stack-machine.
Every opcode has a know effect on the shape of the array on which it
operates. Thus, by knowing the initial shape of each array the algorithm
is able to predict the shape of the resulting array after each opcode
execution. In other words the algorithm is able to predict the shape
of the array on the top of the call stack after each opcode execution.
This is key: every op whose result contains some O(n) element in its
shape has to be executed on the local nodes. On the other hand, every
op whose result is free from O(n) elements is a candidate for execution
on the global node. More information is needed to conclude if indeed
these opcodes are suit for the local or global nodes. This information
comes both from the correctness of merging partial results and from
the height of the stack at this moment These aspects of the algorithm
are not yet implemented, hence it is not technically correct yet.
However, it is correct for linear regression where the first criterion
is enough.

In order to visualize what takes place during the execution on linear
regression we present the opcodes before and after their assignment to
either local or global nodes.

Expression at work for Linear Regression:
E = inv(X.T @ X) @ (X.T @ y)
where X and y are features and target respectively.

This translates to the following instructions:

Initial instructions               | Local node | Global node
===================================|============|============
LOAD X         # (n, k)  BAD       | LOAD X     |
TRANSPOSE      # (k, n)  BAD       | TRANSPOSE  |
LOAD X         # (n, k)  BAD       | LOAD X     |
MATMUL         # (k, k)  GOOD!     | MATMUL  -> | FETCH
LOAD X         # (n, k)  BAD       | LOAD X     |
TRANSPOSE      # (k, n)  BAD       | TRANSPOSE  |
LOAD y         # (n,)    BAD       | LOAD y     |
MATMUL         # (k,)    GOOD!     | MATMUL  -> | FETCH
MATMUL         # (k,)    GOOD!     |            | MATMUL

where BAD/GOOD refers to the fact that n is present/not present in the
shape of the top-of-the-stack element. The FETCH instruction here
represents a result that is fetched from the local nodes and
pushed onto the stack.

Future work:
-Implement additional criteria for breaking opcodes into local/global
components coming from merging strategy correctness and from stack
height.
-Apply to logistic regression.